poc_kokkos_rs

Module routines

parallel statement related code

This module contains code used for the implementation of parallel statements, e.g. parallel_for, a Kokkos specific implementation of commonly used patterns.

Parameters of aforementionned statements are defined in the parameters sub-module.

Dispatch code is defined in the dispatch sub-module.

Currently implemented statements:

parallel_for

Modules

dispatch

kernel dispatch code

parameters

parallel statement parameterization code

Enums

StatementError

Enum used to classify possible errors occuring in a parallel statement.

Functions

parallel_for

Parallel For statement.