poc_kokkos_rs::routines

Module dispatch

kernel dispatch code

This module contains all code used to dispatch computational kernels onto specified devices. Note that the documentation is feature-specific when the items are, i.e. documentation is altered by enabled features.

The methods desccribed in this module are not meant to be used directly, they are only building blocks for the parallel statements.

Enums

DispatchError

Enum used to classify possible dispatch errors.

Functions

cpu

CPU dispatch routine of for statements. Implementation depends on enabled feature(s).

gpu

GPU Dispatch routine of for statements. UNIMPLEMENTED

serial

CPU dispatch routine of for statements. Does not depend on enabled feature(s).