poc_kokkos_rs::routines::dispatch

Function cpu

pub fn cpu<const N: usize>(
    execp: ExecutionPolicy<N>,
    kernel: SerialForKernelType<'_, N>,
) -> Result<(), DispatchError>

CPU dispatch routine of for statements. Implementation depends on enabled feature(s).

The dispatch function execute the kernel accordingly to the directives contained in the execution policy. The kernel signature varies according to enabled features.

Possible Kernel Signatures

rayon feature enabled: ForKernelType

threads feature enabled: Box<impl Fn(KernelArgs<N>) + Send + Sync + 'a + Clone>

no feature enabled: fall back to SerialForKernelType

The threads implementation cannot currently use the generic ForKernelType because of the Clone requirement.

Current version: no feature