poc_kokkos_rs::routines

Module parameters

parallel statement parameterization code

This module contains all code used to parameterize parallel statements. Some parameters are direct Kokkos replicates while others are Rust-specific.

Notable differences include:

ExecutionPolicy struct: Instead of having multiple types of execution policies for each range, with re-occuring parameters, range specification is now a subparameter of execution policies.

Structs

ExecutionPolicy

Execution Policy enum. See Kokkos documentation for explanation on their model.

Enums

ExecutionSpace

Execution Space enum.

RangePolicy

Range Policy enum.

Schedule

Scheduling enum. CURRENTLY IGNORED.