poc_kokkos_rs::view

Module parameters

view parameterization code

This module contains all code used to parameterize views. Some parameters are direct Kokkos replicates while others are Rust-specific. Currently supported parameters include:

Type of the data owned by the view (Rust-specific)

Memory layout

Possible future implementations include:

Memory space

Memory traits

Enums

DataType

Enum used to identify the type of data the view is holding.

Layout

Enum used to represent data layout. Struct enums is used in order to increase readability.

Constants

MAX_VIEW_DEPTH

Maximum possible depth (i.e. number of dimensions) for a view.

Traits

DataTraits

Supertrait with common trait that elements of a View should implement.

Functions

compute_stride

Compute correct strides of each index using dimensions and specified layout.

Type Aliases

InnerDataType

Generic alias for elements of type T of a View.