poc_kokkos_rs::view::parameters

Type Alias InnerDataType

pub type InnerDataType<T> = T;

Generic alias for elements of type T of a View.

This alias automatically changes according to features to ensure thread-safety of Views. There are two possible values:

any feature enabled: InnerDataType<T> = Atomic<T>. By adding the atomic wrapping, operations on views can be implemented using thread-safe methods.

no feature enabled: InnerDataType<T> = T.

Current version: no feature