poc_kokkos_rs::view

Type Alias ViewRW

pub type ViewRW<'a, const N: usize, T> = ViewBase<'a, N, T>;

View type owning a mutable borrow to the data it yields access to, i.e. a read-write mirror.

Aliased Type

pub struct ViewRW<'a, const N: usize, T> {
    pub data: DataType<'a, T>,
    pub layout: Layout<N>,
    pub dim: [usize; N],
    pub stride: [usize; N],
}

Fields
data: DataType<'a, T>

Data container. Depending on the type, it can be a vector (Owned), a reference (ReadOnly) or a mutable reference (ReadWrite).

layout: Layout<N>

Memory layout of the view. Refer to Kokkos documentation for more information.

dim: [usize; N]

Dimensions of the data represented by the view. The view can:

be a vector (1 dimension)

be a multi-dimensionnal array (up to 8 dimensions)

The number of dimensions is referred to as the depth. Dimension 0, i.e. scalar, is not directly supported.

stride: [usize; N]

Stride between each element of a given dimension. Computed automatically for Layout::Left and Layout::Right.