poc_kokkos_rs::functor

Type Alias ForKernelType

pub type ForKernelType<'a, const N: usize> = SerialForKernelType<'a, N>;

parallel_for kernel type. Depends on enabled feature(s).

This type alias is configured according to enabled feature in order to adjust the signatures of kernels to match the requirements of the underlying dispatch routines.

Possible Values

rayon feature enabled: Box<dyn Fn(KernelArgs<N>) + Send + Sync + 'a>

threads feature enabled: Box<dyn Fn(KernelArgs<N>) + Send + 'a>

no feature enabled: fall back to SerialForKernelType

Current version: no feature

Aliased Type

pub struct ForKernelType<'a, const N: usize>(/* private fields */);